Make sure the internal link is a comment before following it
Besides links to other comments, meneame also allows adding
links to tags or users. This makes sure that we've clicked on
a link to a comment before following it. Still, it's very difficult
for an user to click on a link: it has to be done in a different way.